Move lens pkg to internal dir
