+++ internal/lens/registry.go
